MON-30: move to node v0.6 and restify 1.x (attempt to fix 'npm WARN ...  skipping, cannot run ...')
The problem stems from adding the 'dev' dep on the 'all' target
recently. This resulted in 'npm install && npm link ./common' at
the top-level before the 'common' target was run. That 'npm link
./common' then tried to be helpful and do the build in ./common.
As 'npm help link' says:

    `npm link ../node-redis` is the equivalent of doing:

         (cd ../node-redis; npm link)
         npm link redis

However that isn't quite equivalent. That 'npm link' actually
runs with the live code `npm.dir` set to the *source* dir's
'./node_modules' dir. If you are also running as root (as
we are for the jenkins build) then you fall afoul of
this check:
<https://github.com/isaacs/npm/blob/f322b0d00a/lib/utils/lifecycle.js#L42-46>

Solutions are either:
- build in the 'common' dir first (i.e. `cd common && npm install`),
  which we'll do here
- run npm with '--unsafe-perms', which we'll NOT do here, but is
  probably a good thing -- judging only because this safe perms
  attempted favour in npm seems to bite frequently