MON-113: maintenance window support (prep work: agent/machine field refactoring)
Transition from machine/server to agent/machine fields on probes.
All probes will have an 'agent' UUID that identifies where this probe is
to run. For "runLocally" probe types (e.g. logscan), the 'agent' and
'machine' are the same and either is allowed to be specified ('agent'
would be preferred but 'machine' is more logical to explain to a user
for those types).

For remote probes (e.g. icmp, http) a 'machine' field is optional and
advisory only -- e.g. an http check might be hitting a load balancer so
isn't necessarily tied to one machine.

Changed APIs:
    sdc-amon /agentprobes?agent=UUID
        Query param is always 'agent' now.
    sdc-amon /pub/:login/probe/:probe -X PUT ...
        'machine/server' arg is 'agent' or 'machine' (see above)

Other changes:
- Also refactor some internal usage of "global" in var names to "vmhost"
  and "local" -> "vmguest" to clarify the scope.
- Some specing out a larger monitoring use case (will be added to docs)
- Clarify more some of the fields on Amon event objects.
- Planning out the maintenance windows facility.