Revert "MON-240: Amon config file confusion (wait on config-agent)"
This reverts commit 242627899423d3410d3d3252103127b8855836c5.
This dep isn't necessary. The actual *setup* time creation of the
amon-master config file is guaranteed because mdata:execute
calls 'config-agent/agents.js -s' to synchronously write out the
manifests for this instacne (e.g. the amon-msater config file) and
then *later* calls the zone's "setup" and "configure" scripts which
import and enable to amon-master svc.