MON-252: amon hammers UFDS with lots of zones in play (less frequent amon-relay polling)
Drop amon-relay polling interval from 30s to 30 *minutes*,
unfortunately. Also, ensure that an amon-relay poll for 100 VMs does
not do 100 request as fast as possible. It will space them out with
50ms-500ms breaks.

Also switch to setTimeout instead of setInterval
for rescheduling polls. A theory on multiple polls for a single
VM UUID in east-1 data when it was hurting was that multiple
intervals were running simultaneously.